+++ zerver/tests/test_openapi.py
@@ -1064,6 +1064,7 @@ class OpenAPIRegexTest(ZulipTestCase):
             == "/users/{user_id}/subscriptions/{stream_id}"
         )
         assert find_openapi_endpoint("/users/iago@zulip.com/presence") == "/users/{email}/presence"
+        assert find_openapi_endpoint("/users/iago@zulip.com") == "/users/{email}"
         assert find_openapi_endpoint("/messages/23") == "/messages/{message_id}"
         assert find_openapi_endpoint("/realm/emoji/realm_emoji_1") == "/realm/emoji/{emoji_name}"